+++ src/Divverence.MarbleTesting.Akka/Async/IWaitForIdleOrActive.cs
@@ -2,7 +2,7 @@ using System.Threading.Tasks;
 
 namespace Divverence.MarbleTesting.Akka.Async
 {
-    public interface IWaitForIdleOrActive
+    internal interface IWaitForIdleOrActive
     {
         Task Active();
         Task Idle();